don't always remove lib folder

+++ install.sh
@@ -31,7 +31,7 @@ if [ $SDL_MIXER ] && [ ! -f "./lib/libSDL2_mixer.so" ]; then
 fi
 
 # $OUT_DIR is provided to us by cargo
-if [ -d ./lib/ ]; then
+if [ -d ./lib/ ] && [ $OUT_DIR ]; then
     cp -r ./lib/* $OUT_DIR
     rm -rf ./lib
 fi
